requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
Pillow>=10.0
python-slugify>=8.0
python-dotenv>=1.0
